# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API de controle de consumo real de combustível. Todas as rotas sob /api exigem autenticação via Bearer Token, exceto /api/auth/login.",
        "title": "API Consumo Real",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/abastecimentos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista os abastecimentos, podendo aplicar filtros.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Abastecimentos"
                ],
                "summary": "Listar Abastecimentos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da empresa",
                        "name": "empresa_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID do local",
                        "name": "local_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID da bomba",
                        "name": "bomba_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID do patrimônio",
                        "name": "patrimonio_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID do frentista",
                        "name": "frentista_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID do combustível",
                        "name": "combustivel_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tipo (ABASTECIMENTO, TRANSFERENCIA)",
                        "name": "tipo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/consumo-real-server_internal_domain_abastecimento.Abastecimento"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra um novo abastecimento de combustível.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Abastecimentos"
                ],
                "summary": "Cadastrar Abastecimento",
                "parameters": [
                    {
                        "description": "Dados do abastecimento",
                        "name": "abastecimento",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_routes.abastecimentoRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_domain_abastecimento.Abastecimento"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/abastecimentos/transferencias": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra uma transferência de combustível entre reservatórios.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Abastecimentos"
                ],
                "summary": "Registrar transferência",
                "parameters": [
                    {
                        "description": "Dados da transferência",
                        "name": "transferencia",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_routes.transferenciaRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_domain_abastecimento.Abastecimento"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/abastecimentos/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os dados completos de um abastecimento.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Abastecimentos"
                ],
                "summary": "Buscar Abastecimento por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do abastecimento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_domain_abastecimento.Abastecimento"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza os dados de um abastecimento.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Abastecimentos"
                ],
                "summary": "Atualizar Abastecimento",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do abastecimento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados do abastecimento",
                        "name": "abastecimento",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_routes.abastecimentoUpdateRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_domain_abastecimento.Abastecimento"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove um abastecimento do sistema.",
                "tags": [
                    "Abastecimentos"
                ],
                "summary": "Excluir Abastecimento",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do abastecimento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Abastecimento excluído com sucesso"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auditorias": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista as movimentações registradas no sistema. Usuários com empresa veem\napenas a própria empresa; usuários sem empresa (ADMIN_BASE) acessam todas\ne devem informar o parâmetro empresa_id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auditoria"
                ],
                "summary": "Listar auditorias",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da empresa (obrigatório para ADMIN_BASE)",
                        "name": "empresa_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nome da entidade (ex.: empresas, usuarios)",
                        "name": "entidade",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tipo de operação (CREATE, UPDATE, DELETE)",
                        "name": "operacao",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID do usuário responsável",
                        "name": "usuario_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Início do período (RFC3339, ex.: 2026-08-01T00:00:00-03:00)",
                        "name": "de",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fim do período (RFC3339, ex.: 2026-08-18T23:59:59-03:00)",
                        "name": "ate",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limite de registros por página (padrão 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Deslocamento para paginação",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/consumo-real-server_internal_domain_auditoria.Auditoria"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Autentica o usuário com e-mail e senha, retornando um token JWT de acesso.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autenticar usuário",
                "parameters": [
                    {
                        "description": "Credenciais de acesso",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_routes.loginRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_application_usuario.LoginResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os dados do usuário autenticado com base no token JWT.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Usuário autenticado",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_domain_usuario.Usuario"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "consumo-real-server_internal_shared_apperror.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/consumo-real-server_internal_shared_apperror.ErrorBody"
                }
            }
        },
...
        "internal_routes.abastecimentoRequestBody": {
            "type": "object",
            "properties": {
                "bico_id": {
                    "type": "integer"
                },
                "bomba_id": {
                    "type": "integer"
                },
                "empresa_id": {
                    "type": "integer"
                },
                "frentista_id": {
                    "type": "integer"
                },
                "horimetro": {
                    "type": "number"
                },
                "local_id": {
                    "type": "integer"
                },
                "odometro": {
                    "type": "number"
                },
                "patrimonio_id": {
                    "type": "integer"
                },
                "preco_unitario": {
                    "type": "number"
                },
                "quantidade": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Cole apenas o token JWT (o prefixo \"Bearer \" é adicionado automaticamente).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}